clean and refine softmax and gemm benchmarks
