+++ .github/workflows/triton-benchmarks.yml
@@ -101,7 +101,8 @@ jobs:
         run: |
           source /home/runner/intel/oneapi/mkl/2024.1/env/vars.sh
           cd benchmarks
-          python setup.py install
+          python setup.py bdist_wheel
+          pip install python/dist/*.whl
 
       - name: Run xetla benchmark
         run: |
