--- benchmark/triton_intel_benchmark/__init__.py
@@ -1,3 +1,0 @@
-import torch
-import intel_extension_for_pytorch
-from . import xetla_kernel
