--- benchmark/xetla_kernel/softmax/CMakeLists.txt
@@ -1,1 +1,0 @@
-target_include_directories(xetla_kernel PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
