--- benchmark/xetla_kernel/xetla-library.conf
@@ -1,1 +1,0 @@
-ae46a690bac364a93437e248418636c2a8423134
