+++ benchmarks/xetla_benchmark/fused_softmax.py
@@ -38,6 +38,31 @@ def naive_softmax(x):
     return ret
 
 
+@triton.autotune(
+    configs=[
+        triton.Config({'BLOCK_SIZE': 128}, num_warps=32),
+        triton.Config({'BLOCK_SIZE': 256}, num_warps=32),
+        triton.Config({'BLOCK_SIZE': 512}, num_warps=32),
+        triton.Config({'BLOCK_SIZE': 1024}, num_warps=32),
+        triton.Config({'BLOCK_SIZE': 2048}, num_warps=32),
+        triton.Config({'BLOCK_SIZE': 128}, num_warps=16),
+        triton.Config({'BLOCK_SIZE': 256}, num_warps=16),
+        triton.Config({'BLOCK_SIZE': 512}, num_warps=16),
+        triton.Config({'BLOCK_SIZE': 1024}, num_warps=16),
+        triton.Config({'BLOCK_SIZE': 2048}, num_warps=16),
+        triton.Config({'BLOCK_SIZE': 128}, num_warps=8),
+        triton.Config({'BLOCK_SIZE': 256}, num_warps=8),
+        triton.Config({'BLOCK_SIZE': 512}, num_warps=8),
+        triton.Config({'BLOCK_SIZE': 1024}, num_warps=8),
+        triton.Config({'BLOCK_SIZE': 2048}, num_warps=8),
+        triton.Config({'BLOCK_SIZE': 128}, num_warps=4),
+        triton.Config({'BLOCK_SIZE': 256}, num_warps=4),
+        triton.Config({'BLOCK_SIZE': 512}, num_warps=4),
+        triton.Config({'BLOCK_SIZE': 1024}, num_warps=4),
+        triton.Config({'BLOCK_SIZE': 2048}, num_warps=4),
+    ],
+    key=[],
+)
 @triton.jit
 def softmax_kernel(output_ptr, input_ptr, input_row_stride, output_row_stride, n_cols, BLOCK_SIZE: tl.constexpr):
     # The rows of the softmax are independent, so we parallelize across those
@@ -70,23 +95,19 @@ def softmax(x):
     # increasing the number of warps (`num_warps`) over which each row is distributed.
     # You will see in the next tutorial how to auto-tune this value in a more natural
     # way so you don't have to come up with manual heuristics yourself.
-    num_warps = 4
-    if BLOCK_SIZE >= 2048:
-        num_warps = 8
-    if BLOCK_SIZE >= 4096:
-        num_warps = 16
+    num_warps = 32
+    # if BLOCK_SIZE >= 2048:
+    #     num_warps = 8
+    # if BLOCK_SIZE >= 4096:
+    #     num_warps = 16
     # Allocate output
     y = torch.empty_like(x)
     # Enqueue kernel. The 1D launch grid is simple: we have one kernel instance per row o
     # f the input matrix
     softmax_kernel[(n_rows, )](
-        y,
-        x,
-        x.stride(0),
-        y.stride(0),
-        n_cols,
-        num_warps=num_warps,
-        BLOCK_SIZE=BLOCK_SIZE,
+        y, x, x.stride(0), y.stride(0), n_cols
+        # num_warps=num_warps,
+        # BLOCK_SIZE=BLOCK_SIZE,
     )
     return y
 
@@ -106,7 +127,7 @@ def softmax(x):
             "Triton",
             "Torch (native)",
             "Torch (jit)",
-            "xetla",
+            "Xetla",
         ],  # label name for the lines
         styles=[('blue', '-'), ('green', '-'), ('green', '--'), ('black', ':')],  # line styles
         ylabel="GB/s",  # label name for the y-axis
@@ -114,18 +135,19 @@ def softmax(x):
         args={'M': 4096},  # values for function arguments not in `x_names` and `y_name`
     ))
 def benchmark(M, N, provider):
-    x = torch.randn(M, N, device='xpu', dtype=torch.float32)
+    x = torch.randn(M, N, device='xpu', dtype=torch.bfloat16)
     quantiles = [0.5, 0.2, 0.8]
     if provider == 'torch-native':
-        ms, min_ms, max_ms = triton.testing.do_bench(lambda: torch.softmax(x, axis=-1), quantiles=quantiles)
+        ms, min_ms, max_ms = triton.testing.do_bench(lambda: torch.softmax(x, axis=-1), quantiles=quantiles, warmup=10,
+                                                     rep=10)
     if provider == 'triton':
-        ms, min_ms, max_ms = triton.testing.do_bench(lambda: softmax(x), quantiles=quantiles)
+        ms, min_ms, max_ms = triton.testing.do_bench(lambda: softmax(x), quantiles=quantiles, warmup=10, rep=10)
     if provider == 'torch-jit':
-        ms, min_ms, max_ms = triton.testing.do_bench(lambda: naive_softmax(x), quantiles=quantiles)
+        ms, min_ms, max_ms = triton.testing.do_bench(lambda: naive_softmax(x), quantiles=quantiles, warmup=10, rep=10)
     if provider == 'xetla':
         name = "softmax_shape_{}_{}".format(N, N)
         func = getattr(xetla_kernel, name)
-        ms, min_ms, max_ms = triton.testing.do_bench(lambda: func(x, 0), quantiles=quantiles)
+        ms, min_ms, max_ms = triton.testing.do_bench(lambda: func(x, 0), quantiles=quantiles, warmup=10, rep=10)
     gbps = lambda ms: 2 * x.nelement() * x.element_size() * 1e-9 / (ms * 1e-3)
     return gbps(ms), gbps(max_ms), gbps(min_ms)
