+++ benchmarks/xetla_kernel/softmax/softmax_config.hpp
@@ -13,41 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-#ifndef TRITON_MICROBENCH_SOFTMAKX_CONFIG_H
-#define TRITON_MICROBENCH_SOFTMAKX_CONFIG_H
+#ifndef TRITONBENCHMARK_TEST_H
+#define TRITONBENCHMARK_TEST_H
 
 #include <sycl.hpp>
-
-class mat0_96x2048x2048_bf16 {
-public:
-  static constexpr size_t mat_n = 2048;
-  static constexpr size_t mat_m = 2048 * 96;
-  static constexpr size_t wg_n = mat_n;
-  static constexpr size_t wg_m = 4;
-  static constexpr size_t sg_n = 512;
-  static constexpr size_t sg_m = 4;
-  using data_type_in = sycl::ext::oneapi::bfloat16;
-  using data_type_out = sycl::ext::oneapi::bfloat16;
-  using data_type_acc = float;
-};
-
-class mat1_96x2048x2048_bf16 {
-public:
-  static constexpr size_t mat_n = 2048;
-  static constexpr size_t mat_m = 2048 * 96;
-  static constexpr size_t wg_n = mat_n;
-  static constexpr size_t wg_m = 1;
-  static constexpr size_t sg_n = 2048;
-  static constexpr size_t sg_m = 1;
-  using data_type_in = sycl::ext::oneapi::bfloat16;
-  using data_type_out = sycl::ext::oneapi::bfloat16;
-  using data_type_acc = float;
-};
 
 class mat1_256x256_bf16_cfg0 {
 public:
   static constexpr size_t mat_n = 256;
-  static constexpr size_t mat_m = 4096;
+  static constexpr size_t mat_m = 256;
   static constexpr size_t wg_n = mat_n;
   static constexpr size_t wg_m = 4; // 1 4 8 16
   static constexpr size_t sg_n = mat_n;
@@ -60,7 +34,7 @@ public:
 class mat1_1024x1024_bf16_cfg0 {
 public:
   static constexpr size_t mat_n = 1024;
-  static constexpr size_t mat_m = 4096;
+  static constexpr size_t mat_m = 1024;
   static constexpr size_t wg_n = mat_n;
   static constexpr size_t wg_m = 4; // 1 4 8 16
   static constexpr size_t sg_n = mat_n;
@@ -73,7 +47,7 @@ public:
 class mat1_2048x2048_bf16_cfg0 {
 public:
   static constexpr size_t mat_n = 2048;
-  static constexpr size_t mat_m = 4096;
+  static constexpr size_t mat_m = 2048;
   static constexpr size_t wg_n = mat_n;
   static constexpr size_t wg_m = 4; // 1 4 8 16
   static constexpr size_t sg_n = mat_n;
@@ -87,13 +61,13 @@ class mat1_4096x4096_bf16_cfg0 {
 public:
   static constexpr size_t mat_n = 4096;
   static constexpr size_t mat_m = 4096;
-  static constexpr size_t wg_n = mat_n;
+  static constexpr size_t wg_n = mat_n / 2;
   static constexpr size_t wg_m = 4; // 1 4 8 16
-  static constexpr size_t sg_n = mat_n;
+  static constexpr size_t sg_n = mat_n / 2;
   static constexpr size_t sg_m = 1;
   using data_type_in = sycl::ext::oneapi::bfloat16;
   using data_type_out = sycl::ext::oneapi::bfloat16;
   using data_type_acc = float;
 };
 
-#endif // TRITON_MICROBENCH_SOFTMAKX_CONFIG_H
+#endif // TRITONBENCHMARK_TEST_H
