+++ python/triton/testing.py
@@ -149,6 +149,7 @@ def do_bench(fn, warmup=25, rep=100, grad_to_none=None, quantiles=None, fast_flu
                 x.grad = None
         # we clear the L2 cache before each run
         cache.zero_()
+        synchronize()
         # record time of `fn`
         start_times[i] = datetime.now()
         fn()
